requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai==2.8.1
requests==2.32.3
python-multipart==0.0.12
pydantic==2.9.0
python-dotenv==1.0.1
prometheus-client==0.21.0
google-generativeai==0.8.3
google-genai==1.30.0
Pillow==10.4.0
httpx>=0.28.1
